Fix too low spacing between counter and button
If the text next to the counter values is longer than usual (e.g.
German translation), the spacing between the counter and the button
below gets too low which does look somewhat broken to me.
Adding a bottom margin to the table containing the counter values and
text solves this issue despite pushing the header a bit down.